Clean up debug-log parser visibility, file layout, and handler naming
