+++ src/EventLogExpert.UI/DebugLog/DebugLogModal.razor
@@ -79,7 +79,7 @@
             </span>
             <div class="debug-log-footer-actions debug-log-footer-right">
                 <Button OnClick="Refresh">Refresh</Button>
-                <Button OnClick="ClearLog">Clear Log</Button>
+                <Button OnClick="ClearDebugLogFileAsync">Clear Log</Button>
             </div>
         </div>
     </ExtraFooterContent>
